grpc_version/requirements.txt
#add dependenies here
grpcio-tools>=1.19.0
grpcio>=1.19.0
